chore: MyPageMain 코드 분리
